Linux rendering and compilation fixes

+++ Core/Code/DataManagement/mitkSurfaceCutterCGAL.cpp
@@ -124,7 +124,7 @@ public:
         pointsOut->Allocate(2 * segments.size());
         pdOut->Allocate(segments.size());
 
-        for (int i = 0; i < segments.size(); ++i) {
+        for (size_t i = 0; i < segments.size(); ++i) {
             CGAL_Segment* s = boost::get<CGAL_Segment>(&segments[i].first);
 
             if (!s) {
